+++ Sources/WebSocket/Frame+Processing.swift
@@ -1,18 +1,24 @@
 extension WebSocket {
+    /// A helper that processes a frame and directs it to the proper handler.
+    ///
+    /// Automatically replies to `ping` and automatically handles `close`
     func processFrame(_ frame: Frame) {
+        // Unmasks the data so it's readable
         frame.unmask()
         
         func processString() {
-            // invalid string
+            // If this is an UTF-8 invalid string
             guard let string = frame.payload.string() else {
                 self.connection.client.close()
                 return
             }
             
+            // Stream to the textStream's listener
             self.textStream.outputStream?(string)
         }
         
         func processBinary() {
+            // Stream to the binaryStream's listener
             self.binaryStream.outputStream?(frame.payload)
         }
